Reorganize component tests into pixels subdirectory
